Pulling main to dev

+++ src/Howbot.Infrastructure/Data/EfRepository.cs
@@ -20,7 +20,7 @@ public class EfRepository : IRepository
     _dbContext = dbContext;
   }
 
-  public T GetById<T>(int id) where T : BaseEntity
+  public T GetById<T>(ulong id) where T : BaseEntity
   {
     return _dbContext.Set<T>().SingleOrDefault(e => e.Id == id);
   }
